+++ src/lagrangian_node.rs
@@ -2,6 +2,6 @@ use crate::common;
 
 struct LagrangianNode {
     pos: common::Vec3f,
-    area: f32,
+    area: f64,
     normal: common::Vec3f,
 }
