+++ src/stl_triangle.rs
@@ -17,17 +17,17 @@ impl TriangleSTL {
         point2: common::Vec3f,
         normal: common::Vec3f,
     ) -> TriangleSTL {
-        assert_almost_equal!(&normal.norm(), 1.0, 1e-5f32);
+        assert_almost_equal!(&normal.norm(), 1.0, 1e-5f64);
         return TriangleSTL { point0, point1, point2, normal };
     }
 
-    pub fn normalize(&mut self, factor: f32, offset: common::Vec3f) {
+    pub fn normalize(&mut self, factor: f64, offset: common::Vec3f) {
         self.point0.transform(factor, offset);
         self.point1.transform(factor, offset);
         self.point2.transform(factor, offset);
     }
 
-    pub fn area(self) -> f32 {
+    pub fn area(self) -> f64 {
         let tr_vec1 = self.point0 - self.point1;
         let tr_vec2 = self.point2 - self.point1;
         let cross_prod: common::Vec3f = tr_vec1.cross(tr_vec2);
@@ -57,17 +57,17 @@ impl fmt::Display for TriangleSTL {
     }
 }
 
-fn get_factor_offset(min_vals: common::Vec3f, max_vals: common::Vec3f) -> (f32, common::Vec3f) {
+fn get_factor_offset(min_vals: common::Vec3f, max_vals: common::Vec3f) -> (f64, common::Vec3f) {
     // Params are: the minimal value, and the difference between min_max for x
     // These can be used to normalize points
-    let div_factor = (max_vals.x - min_vals.x) * 0.01f32; // normalize between 0 and 100
+    let div_factor = (max_vals.x - min_vals.x) * 0.01f64; // normalize between 0 and 100
     let offset = common::Vec3f { x: min_vals.x, y: min_vals.y, z: min_vals.z };
     return (div_factor, offset);
 }
 
 fn get_triangles_min_max(triangles: &Vec<TriangleSTL>) -> (common::Vec3f, common::Vec3f) {
-    let mut min_vals = common::Vec3f { x: f32::MAX, y: f32::MAX, z: f32::MAX };
-    let mut max_vals = common::Vec3f { x: f32::MIN, y: f32::MIN, z: f32::MIN };
+    let mut min_vals = common::Vec3f { x: f64::MAX, y: f64::MAX, z: f64::MAX };
+    let mut max_vals = common::Vec3f { x: f64::MIN, y: f64::MIN, z: f64::MIN };
     for t in triangles {
         for p in [t.point0, t.point1, t.point2] {
             if p.x < min_vals.x {
